# backend/src/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "SysMap Social Network - OpenAPI 3.0",
    "description": "Backend of the Social Network project of SysMap fullstack bootcamp.",
    "contact": {
      "email": "kellyplcastelo@gmail.com"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:4000/v1"
    }
  ],
...
  "paths": {
    "/posts": {
      "get": {
        "tags": ["posts"],
        "summary": "Get all existing posts",
        "description": "Get all existing posts",
        "operationId": "getPosts",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostResponse"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHORIZED"
          }
        },
        "security": [
          {
            "BearerAuth": {}
          }
        ]
      },
      "post": {
        "tags": ["posts"],
        "summary": "Creates a new post",
        "description": "Creates a new post",
        "operationId": "createNewPost",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "CREATED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostResponse"
                }
              }
            }
          },
          "400": {
            "description": "BAD REQUEST",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorArrayResponse"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHORIZED"
          }
        },
        "security": [
          {
            "BearerAuth": {}
          }
        ]
      }
    },
    "/posts/{id}": {
      "get": {
        "tags": ["posts"],
        "summary": "Find posts by id",
        "description": "Find posts by id",
        "operationId": "findPostById",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "uuid of the post to get"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostResponse"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHORIZED"
          },
          "404": {
            "description": "NOT FOUND"
          }
        },
        "security": [
          {
            "BearerAuth": {}
          }
        ]
      },
      "put": {
        "tags": ["posts"],
        "summary": "Update a post",
        "description": "This can only be done by the logged in user.",
        "operationId": "updatePost",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "uuid of the post to update"
          }
        ],
        "requestBody": {
          "description": "Update an existent post",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostRequest"
              }
            }
          }
        },
        "responses": {
          "203": {
            "description": "UNDOCUMENTED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostResponse"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHORIZED"
          },
          "404": {
            "description": "NOT FOUND"
          }
        },
        "security": [
          {
            "BearerAuth": {}
          }
        ]
      },
      "delete": {
        "tags": ["posts"],
        "summary": "Deletes a post",
        "description": "delete a post",
        "operationId": "deletePost",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "uuid of the post to delete"
          }
        ],
        "responses": {
          "204": {
            "description": "NO CONTENT"
          },
          "401": {
            "description": "UNAUTHORIZED"
          },
          "404": {
            "description": "NOT FOUND"
          }
        },
        "security": [
          {
            "BearerAuth": {}
          }
        ]
      }
    },
    "/posts/{id}/like": {
      "post": {
        "tags": ["posts"],
        "summary": "Likes the post",
        "description": "Likes the post, liking the same post twice will result in a bad request",
        "operationId": "likesPost",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "uuid of the post to like"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostResponseLikeUnlike"
                }
              }
            }
          },
          "400": {
            "description": "BAD REQUEST"
          },
          "401": {
            "description": "UNAUTHORIZED"
          },
          "404": {
            "description": "NOT FOUND"
          }
        },
        "security": [
          {
            "BearerAuth": {}
          }
        ]
      }
    },
    "/posts/{id}/unlike": {
      "post": {
        "tags": ["posts"],
        "summary": "Unlikes the post",
        "description": "Unlikes the post, unliking the same post twice will result in a bad request",
        "operationId": "unlikesPost",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "uuid of the post to like"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostResponseLikeUnlike"
                }
              }
            }
          },
          "400": {
            "description": "BAD REQUEST"
          },
          "401": {
            "description": "UNAUTHORIZED"
          },
          "404": {
            "description": "NOT FOUND"
          }
        },
        "security": [
          {
            "BearerAuth": {}
          }
        ]
      }
    },
    "/posts/{postId}/comments": {
      "get": {
        "tags": ["comments"],
        "summary": "Get all existing comments from a post",
        "description": "Get all existing posts",
        "operationId": "getPosts",
        "parameters": [
          {
            "in": "path",
            "name": "postId",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "uuid of the post to get comments from"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CommentResponse"
                  }
                }
              }
            }
          },
          "400": {
            "description": "BAD REQUEST",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ErrorArrayResponse"
                  }
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHORIZED"
          },
          "404": {
            "description": "NOT FOUND"
          }
        },
        "security": [
          {
            "BearerAuth": {}
          }
        ]
      },
      "post": {
        "tags": ["comments"],
        "summary": "Create a new comment",
        "description": "Create a new comment",
        "operationId": "createNewComment",
        "parameters": [
          {
            "in": "path",
            "name": "postId",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "uuid of the post to get comments from"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommentRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "CREATED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommentResponse"
                }
              }
            }
          },
          "400": {
            "description": "BAD REQUEST",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ErrorArrayResponse"
                  }
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHORIZED"
          },
          "404": {
            "description": "NOT FOUND"
          }
        },
        "security": [
          {
            "BearerAuth": {}
          }
        ]
      }
    },
...
    "/profiles": {
      "get": {
        "tags": ["profile"],
        "summary": "Find all profiles",
        "description": "Find all profiles",
        "operationId": "findProfiles",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProfileResponse"
                  }
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHORIZED"
          }
        },
        "security": [
          {
            "BearerAuth": {}
          }
        ]
      }
    },
    "/profiles/search?q={q}": {
      "get": {
        "tags": ["profile"],
        "summary": "Search for a profile by its name",
        "description": "Search for a profile by its name",
        "operationId": "findProfileByName",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "schema": {
              "type": "string",
              "example": "Kelly Castelo"
            },
            "required": true,
            "description": "profile's name"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileResponse"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHORIZED"
          }
        },
        "security": [
          {
            "BearerAuth": {}
          }
        ]
      }
    },
    "/profile/:id/follow": {
      "post": {
        "tags": ["profile"],
        "summary": "Follow a profile",
        "description": "Follow a profile, returns 400 if you're already following this profile",
        "operationId": "followProfile",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "uuid of the profile to follow"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileResponse"
                }
              }
            }
          },
          "400": {
            "description": "BAD REQUEST"
          },
          "401": {
            "description": "UNAUTHORIZED"
          },
          "404": {
            "description": "NOT FOUND"
          }
        },
        "security": [
          {
            "BearerAuth": {}
          }
        ]
      }
    },
    "/profile/:id/unfollow": {
      "post": {
        "tags": ["profile"],
        "summary": "Unfollow a profile",
        "description": "Unfollow a profile, returns 400 if you're not following this profile",
        "operationId": "unfollowProfile",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "uuid of the profile to unfollow"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileResponse"
                }
              }
            }
          },
          "400": {
            "description": "BAD REQUEST"
          },
          "401": {
            "description": "UNAUTHORIZED"
          },
          "404": {
            "description": "NOT FOUND"
          }
        },
        "security": [
          {
            "BearerAuth": {}
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorArrayResponse": {
        "type": "object",
        "properties": {
          "path": {
            "type": "string",
            "example": "/v1/path/action"
          },
          "status": {
            "type": "integer",
            "example": "400"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "string",
              "example": "Path `user` is required"
            }
          }
        }
      },
...
      "PostResponse": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "637bbc3a69c083c2b0003097"
          },
          "title": {
            "type": "string",
            "example": "To be or not to be"
          },
          "description": {
            "type": "string",
            "example": "That is the question"
          },
          "profile": {
            "$ref": "#/components/schemas/ProfileResponse"
          },
          "comments": {
            "type": "array",
            "items": {
              "type": "string",
              "example": "637bb9d4da4e28ea8606b7c4"
            }
          },
          "likes": {
            "type": "array",
            "items": {
              "type": "string",
              "example": "637bb9d4da4e28ea8606b7c4"
            }
          }
        }
      },
      "PostResponseLikeUnlike": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "637bbc3a69c083c2b0003097"
          },
          "title": {
            "type": "string",
            "example": "To be or not to be"
          },
          "description": {
            "type": "string",
            "example": "That is the question"
          },
          "profile": {
            "type": "string",
            "example": "637bb9d4da4e28ea8606b7c1"
          },
          "comments": {
            "type": "array",
            "items": {
              "type": "string",
              "example": "637bb9d4da4e28ea8606b7c4"
            }
          },
          "likes": {
            "type": "array",
            "items": {
              "type": "string",
              "example": "637bb9d4da4e28ea8606b7c4"
            }
          }
        }
      },
      "PostRequest": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "To be or not to be"
          },
          "description": {
            "type": "string",
            "example": "That is the question"
          }
        },
        "required": ["title", "description"]
      },
      "ProfileResponse": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "637bb9d4da4e28ea8606b7c1"
          },
          "name": {
            "type": "string",
            "example": "Kelly Castelo"
          },
          "user": {
            "type": "string",
            "example": "637bb9d4da4e28ea8606b7bf"
          },
          "following": {
            "type": "array",
            "items": {
              "type": "string",
              "example": "637bb9d4da4e28ea8606b7c7"
            }
          },
          "followers": {
            "type": "array",
            "items": {
              "type": "string",
              "example": "637bb9d4da4e28ea8606b7c8"
            }
          }
        }
      },
...
      "CommentRequest": {
        "type": "object",
        "properties": {
          "description": {
            "type": "string",
            "example": "Amazing post!"
          }
        },
        "required": ["description"]
      },
      "CommentResponse": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "637bbc3a69c083c2b0003097"
          },
          "description": {
            "type": "string",
            "example": "That is the question"
          },
          "profile": {
            "$ref": "#/components/schemas/ProfileResponse"
          },
          "comments": {
            "type": "array",
            "items": {
              "type": "string",
              "example": "637bb9d4da4e28ea8606b7c4"
            }
          },
          "likes": {
            "type": "array",
            "items": {
              "type": "string",
              "example": "637bb9d4da4e28ea8606b7c4"
            }
          }
        }
      }
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}